remove gcc warning

+++ libatalk/cnid/cdb/cnid_cdb_resolve.c
@@ -1,5 +1,5 @@
 /*
- * $Id: cnid_cdb_resolve.c,v 1.2 2005-04-28 20:49:59 bfernhomberg Exp $
+ * $Id: cnid_cdb_resolve.c,v 1.3 2005-05-03 14:55:13 didg Exp $
  */
 
 #ifdef HAVE_CONFIG_H
@@ -10,7 +10,7 @@
 #include "cnid_cdb_private.h"
 
 /* Return the did/name pair corresponding to a CNID. */
-char *cnid_cdb_resolve(struct _cnid_db *cdb, cnid_t *id, void *buffer, u_int32_t len) {
+char *cnid_cdb_resolve(struct _cnid_db *cdb, cnid_t *id, void *buffer, size_t len) {
     CNID_private *db;
     DBT key, data;
     int rc;
